[Refactor] Refactor the table source file directory
