+++ src/renderer/views/components/Table/Delete/index.vue
