+++ src/renderer/views/components/Table/Preview/index.vue
